@@ -1,19 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;net471;net472;net48;net5.0</TargetFrameworks>
-    <AssemblyName>BigSolution.Infra.Domain.Tests</AssemblyName>
-    <RootNamespace>BigSolution.Domain</RootNamespace>
-    <Description>Unit testing of domain</Description>
+    <TargetFrameworks>net6.0;net5.0;netcoreapp3.1</TargetFrameworks> <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable> <AssemblyName>BigSolution.Infra.Domain.Tests</AssemblyName>
+    <RootNamespace>BigSolution.Domain</RootNamespace> <Description>Unit testing of domain</Description>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="coverlet.msbuild" Version="3.0.2">
+    <PackageReference Include="coverlet.msbuild" Version="3.1.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="FluentAssertions" Version="5.10.3" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
+    <PackageReference Include="FluentAssertions" Version="6.4.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
     <PackageReference Include="xunit.core" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
       <PrivateAssets>all</PrivateAssets>
