@@ -1,6 +1,6 @@
 ﻿#region Copyright & License
 
-// Copyright © 2020 - 2021 Emmanuel Benitez
+// Copyright © 2020 - 2022 Emmanuel Benitez
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,22 +16,18 @@
 
 #endregion
 
-using System;
 using System.Diagnostics.CodeAnalysis;
-using System.Threading;
-using System.Threading.Tasks;
 
-namespace BigSolution.Domain
+namespace BigSolution.Domain;
+
+[SuppressMessage("ReSharper", "UnusedMember.Global")]
+public interface ITransaction : IDisposable
 {
-    [SuppressMessage("ReSharper", "UnusedMember.Global")]
-    public interface ITransaction : IDisposable
-    {
-        void Commit();
+    void Commit();
 
-        Task CommitAsync(CancellationToken cancellationToken = default);
+    Task CommitAsync(CancellationToken cancellationToken = default);
 
-        void Rollback();
+    void Rollback();
 
-        Task RollbackAsync(CancellationToken cancellationToken = default);
-    }
+    Task RollbackAsync(CancellationToken cancellationToken = default);
 }
